pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xrayctl"
version = "0.1.0"
description = "Custom CLI for JFrog Xray"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31.0",
  "PyYAML>=6.0.1",
]

[project.scripts]
xrayctl = "xrayctl.main:main"
